[FEATURE] Move emptyOption + items to multivalue base class
Arguments emptyOption and items apply to more than simple
select - and Flux simple select cannot be used for relations.

Refactoring these arguments allows flux:field.relation to respect
the already added emptyOption argument.
