+++ Classes/Form/Field/Select.php
@@ -20,46 +20,6 @@ class Select extends AbstractMultiValueFormField
 {
 
     /**
-     * Mixed - string (CSV), Traversable or array of items. Format of key/value
-     * pairs is also optional. For single-dim arrays, key becomes option value
-     * and each member value becomes label. For multidim/Traversable each member
-     * is inspected; if it is a raw value it is used for both value and label,
-     * if it is a scalar value the first item is used as value and the second
-     * as label.
-     *
-     * @var mixed
-     */
-    protected $items = null;
-
-    /**
-     * If not-FALSE, adds one empty option/value pair to the generated selector
-     * box and tries to use this property's value (cast to string) as label.
-     *
-     * @var boolean|string
-     */
-    protected $emptyOption = false;
-
-    /**
-     * If set to TRUE, Flux will attempt to translate the LLL labels of items
-     * provided as CSV values, e.g. items "foo,bar" would try to resolve LLL
-     * values for "LLL:EXT:myext/Resources/Private/Languages/locallang.xlf:foo"
-     * and "LLL:EXT:myext/Resources/Private/Languages/locallang.xlf:bar" to be
-     * used as value labels.
-     *
-     * @var boolean
-     */
-    protected $translateCsvItems = false;
-
-    /**
-     * Special rendering type of this component - supports all values normally
-     * supported by TCA of the "select" field type.
-     *
-     * @var string
-     * @see https://docs.typo3.org/typo3cms/TCAReference/Reference/Columns/Select/Index.html#rendertype
-     */
-    protected $renderType = 'selectSingle';
-
-    /**
      * Displays option icons as table beneath the select.
      *
      * @var boolean
@@ -73,153 +33,8 @@ class Select extends AbstractMultiValueFormField
     public function buildConfiguration()
     {
         $configuration = parent::prepareConfiguration('select');
-        $configuration['items'] = $this->getItems();
         $configuration['showIconTable'] = $this->getShowIconTable();
         return $configuration;
-    }
-
-    /**
-     * @param array $items
-     * @return Select
-     */
-    public function setItems($items)
-    {
-        $this->items = $items;
-        return $this;
-    }
-
-    /**
-     * @return boolean
-     */
-    public function getTranslateCsvItems()
-    {
-        return $this->translateCsvItems;
-    }
-
-    /**
-     * @param boolean $translateCsvItems
-     * @return Select
-     */
-    public function setTranslateCsvItems($translateCsvItems)
-    {
-        $this->translateCsvItems = $translateCsvItems;
-        return $this;
-    }
-
-    /**
-     * @return array
-     */
-    public function getItems()
-    {
-        $items = [];
-        if (true === $this->items instanceof QueryInterface) {
-            $items = $this->addOptionsFromResults($this->items);
-        } elseif (true === is_string($this->items)) {
-            if (false !== strpos($this->items, '..')) {
-                list ($low, $high) = explode('..', $this->items);
-                $itemNames = range($low, $high, 1);
-            } else {
-                $itemNames = GeneralUtility::trimExplode(',', $this->items);
-            }
-            if (!$this->getTranslateCsvItems()) {
-                foreach ($itemNames as $itemName) {
-                    array_push($items, [$itemName, $itemName]);
-                }
-            } else {
-                foreach ($itemNames as $itemName) {
-                    $resolvedLabel = $this->resolveLocalLanguageValueOfLabel(
-                        '',
-                        $this->getPath() . '.option.' . $itemName
-                    );
-                    array_push($items, [$resolvedLabel, $itemName]);
-                }
-            }
-        } elseif (true === is_array($this->items) || true === $this->items instanceof \Traversable) {
-            foreach ($this->items as $itemIndex => $itemValue) {
-                if (true === is_array($itemValue) || true === $itemValue instanceof \ArrayObject) {
-                    array_push($items, $itemValue);
-                } else {
-                    array_push($items, [$itemValue, $itemIndex]);
-                }
-            }
-        }
-        $emptyOption = $this->getEmptyOption();
-        if (false !== $emptyOption) {
-            array_unshift($items, [$emptyOption, '']);
-        }
-        return $items;
-    }
-
-    /**
-     * @param boolean|string $emptyOption
-     * @return Select
-     */
-    public function setEmptyOption($emptyOption)
-    {
-        $this->emptyOption = $emptyOption;
-        return $this;
-    }
-
-    /**
-     * @return boolean|string
-     */
-    public function getEmptyOption()
-    {
-        return $this->emptyOption;
-    }
-
-    /**
-     * @return string
-     */
-    public function getRenderType()
-    {
-        return $this->renderType;
-    }
-
-    /**
-     * @param string $renderType
-     * @return void
-     */
-    public function setRenderType($renderType)
-    {
-        $this->renderType = $renderType;
-    }
-
-    /**
-     * @param QueryInterface $query
-     * @return array
-     */
-    protected function addOptionsFromResults(QueryInterface $query)
-    {
-        $items = [];
-        $results = $query->execute();
-        $type = $query->getType();
-        $table = strtolower(str_replace('\\', '_', $type));
-        $propertyName = $this->getLabelPropertyName($table, $type);
-        foreach ($results as $result) {
-            $uid = $result->getUid();
-            array_push($items, [ObjectAccess::getProperty($result, $propertyName), $uid]);
-        }
-        return $items;
-    }
-
-    /**
-     * @param string $table
-     * @param string $type
-     * @return string
-     */
-    protected function getLabelPropertyName($table, $type)
-    {
-        $typoScript = $this->getConfigurationService()->getAllTypoScript();
-        if (true === isset($typoScript['config']['tx_extbase']['persistence']['classes'][$type])) {
-            $mapping = $typoScript['config']['tx_extbase']['persistence']['classes'][$type];
-            if (true === isset($mapping['mapping']['tableName'])) {
-                $table = $mapping['mapping']['tableName'];
-            }
-        }
-        $labelField = $GLOBALS['TCA'][$table]['ctrl']['label'];
-        $propertyName = GeneralUtility::underscoredToLowerCamelCase($labelField);
-        return $propertyName;
     }
 
     /**
