+++ Classes/ViewHelpers/Field/SelectViewHelper.php
@@ -50,26 +50,11 @@ class SelectViewHelper extends AbstractMultiValueFieldViewHelper
     public function initializeArguments()
     {
         parent::initializeArguments();
-        $this->registerArgument(
+        $this->overrideArgument(
             'items',
             'mixed',
             'Items for the selector; array / CSV / Traversable / Query supported',
             true
-        );
-        $this->registerArgument(
-            'emptyOption',
-            'mixed',
-            'If not-FALSE, adds one empty option/value pair to the generated selector box and tries to use this ' .
-            'property\'s value (cast to string) as label.',
-            false,
-            false
-        );
-        $this->registerArgument(
-            'translateCsvItems',
-            'boolean',
-            'If TRUE, attempts to resolve a LLL label for each value provided as CSV in "items" attribute using ' .
-            'convention for lookup "$field.option.123" if given "123" as CSV item value. Field name is determined ' .
-            'by normal Flux field name conventions'
         );
         $this->registerArgument(
             'renderType',
@@ -96,9 +81,6 @@ class SelectViewHelper extends AbstractMultiValueFieldViewHelper
     {
         /** @var Select $component */
         $component = static::getPreparedComponent('Select', $renderingContext, $arguments);
-        $component->setItems($arguments['items']);
-        $component->setEmptyOption($arguments['emptyOption']);
-        $component->setTranslateCsvItems((boolean) $arguments['translateCsvItems']);
         $component->setRenderType($arguments['renderType']);
         $component->setShowIconTable($arguments['showIconTable']);
         return $component;
